requirements.txt
api==0.0.7
audioread==3.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cmake==3.27.4.1
contourpy==1.1.1
cycler==0.11.0
decorator==5.1.1
filelock==3.12.4
fonttools==4.42.1
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-tools==1.57.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
librosa==0.9.2
lit==16.0.6
llvmlite==0.41.0
mariadb==1.1.8
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
nose==1.3.7
numba==0.58.0
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
Pillow==10.0.1
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.24.2
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
resampy==0.4.2
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
soundfile==0.12.1
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchaudio==2.0.2
torchsummary==1.5.1
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
webrtcvad-wheels==2.0.11.post1
PyJWT